Renamed global to store
